# json/todo/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Ica MinButik - Todo",
    "version" : "1.0.0",
    "title" : "ToDo API",
    "contact" : {
      "email" : "goran.roseen@acando.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/ica-egenkontroll/todo/1.0.0",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/activities" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "searches for activities",
        "description" : "By passing in the appropriate options, you can search for\navailable activity in the system\n",
        "operationId" : "searchActivity",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "$top",
          "in" : "query",
          "description" : "Show only the first n items, see [OData Paging - Top](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374630)",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "$skip",
          "in" : "query",
          "description" : "Skip the first n items, see [OData Paging - Skip](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374631)",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "$filter",
          "in" : "query",
          "description" : "Filter items by property values, see [OData Filtering](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374625), note that expand is not supported",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "$select",
          "in" : "query",
          "description" : "Select properties to be returned, see [OData Select](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374620), note that expand is not supported",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Activity"
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "clients" ],
        "summary" : "adds an Activity",
        "description" : "Adds an activity to the system",
        "operationId" : "addActivity",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "Activity",
          "description" : "Activity to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Activity"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    },
    "/activities/{id}" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "Find activity by Id",
        "description" : "Find an activity by Id",
        "operationId" : "getActivityById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the activity that needs to be fetched",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Activity"
            }
          },
          "400" : {
            "description" : "error"
          },
          "404" : {
            "description" : "Could not find an entry with the given id."
          }
        }
      },
      "put" : {
        "tags" : [ "clients" ],
        "summary" : "updates an activity item",
        "description" : "Updates an activity item in the system",
        "operationId" : "updateActivity",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id of the Activity item to be updated",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "ActivityItem",
          "description" : "Activity item to update",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Activity"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "item updated"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins" ],
        "summary" : "Delete Activity by Id",
        "description" : "Deletes an activity item in the system",
        "operationId" : "deleteActivity",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the activity that needs to be deleted",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Delete successful"
          },
          "400" : {
            "description" : "error"
          },
          "404" : {
            "description" : "Could not find an entry with the given id."
          }
        }
      }
    },
    "/activities/{id}/tags" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "shows for tags belonging to the activity",
        "description" : "By passing in the appropriate options, you can search for\navailable activity in the system\n",
        "operationId" : "searchActivityTag",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the activity that needs to be fetched",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "$top",
          "in" : "query",
          "description" : "Show only the first n items, see [OData Paging - Top](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374630)",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "$skip",
          "in" : "query",
          "description" : "Skip the first n items, see [OData Paging - Skip](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374631)",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "$filter",
          "in" : "query",
          "description" : "Filter items by property values, see [OData Filtering](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374625), note that expand is not supported",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "$select",
          "in" : "query",
          "description" : "Select properties to be returned, see [OData Select](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374620), note that expand is not supported",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Activity"
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "clients" ],
        "summary" : "adds an ActivityTag",
        "description" : "Adds an activityTag to the system",
        "operationId" : "addActivityTag",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the activity that needs to be fetched",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "ActivityTag",
          "description" : "ActivityTag to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Tag"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    },
    "/activities/{id}/tags/{tagId}" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "Find activityTag by tag Id",
        "description" : "Find an activityTag by Tag Id",
        "operationId" : "getActivityTagById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the activity that needs to be fetched",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tagId",
          "in" : "path",
          "description" : "Id of the tag that needs to be fetched",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Tag"
            }
          },
          "400" : {
            "description" : "error"
          },
          "404" : {
            "description" : "Could not find an entry with the given id."
          }
        }
      },
      "delete" : {
        "tags" : [ "admins" ],
        "summary" : "Delete ActivityTag by TagId",
        "description" : "Deletes an activityTag item in the system",
        "operationId" : "deleteActivityTag",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the activity that has a specific tag",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tagId",
          "in" : "path",
          "description" : "Id of the tag that needs to be deleted",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Delete successful"
          },
          "400" : {
            "description" : "error"
          },
          "404" : {
            "description" : "Could not find an entry with the given id."
          }
        }
      }
    },
    "/activities/{id}/instantiateTodo" : {
      "post" : {
        "tags" : [ "clients" ],
        "summary" : "Creates a todo",
        "description" : "Creates a Todo based on the Activity",
        "operationId" : "createTodoFromActivity",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the activity that the todo will be created from",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "todo",
          "description" : "fields needed for instantiation of Todo",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TodoInstantiate"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Tag" : {
      "type" : "object",
      "required" : [ "title" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "123456"
        },
        "title" : {
          "type" : "string",
          "example" : "Rengöring"
        },
        "sortId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "modifiedOn" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-03-14 15:22"
        }
      },
      "example" : {
        "modifiedOn" : "2018-03-14 15:22",
        "sortId" : 1,
        "id" : "123456",
        "title" : "Rengöring"
      }
    },
...
    "Activity" : {
      "type" : "object",
      "required" : [ "activityType", "categoryId", "title" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "123456"
        },
        "title" : {
          "type" : "string",
          "example" : "Städa charken"
        },
        "activityType" : {
          "type" : "string",
          "example" : "basic",
          "description" : "Must be one of \"basic\", \"measure\", \"yesno\""
        },
        "note" : {
          "type" : "string",
          "example" : "Gör så här"
        },
        "contentUrl" : {
          "type" : "string",
          "format" : "url",
          "example" : "https://episerver.ica.se/getContent?id=4711"
        },
        "createdAt" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-03-14 15:22"
        },
        "priority" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "categoryId" : {
          "type" : "string",
          "example" : "123"
        },
        "categoryTitle" : {
          "type" : "string",
          "example" : "HACCP"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Tag"
          }
        },
        "modifiedOn" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-03-14 15:22"
        }
      },
      "example" : {
        "note" : "Gör så här",
        "createdAt" : "2018-03-14 15:22",
        "contentUrl" : "https://episerver.ica.se/getContent?id=4711",
        "modifiedOn" : "2018-03-14 15:22",
        "id" : "123456",
        "title" : "Städa charken",
        "activityType" : "basic",
        "priority" : 1,
        "categoryId" : "123",
        "categoryTitle" : "HACCP",
        "tags" : [ {
          "modifiedOn" : "2018-03-14 15:22",
          "sortId" : 1,
          "id" : "123456",
          "title" : "Rengöring"
        }, {
          "modifiedOn" : "2018-03-14 15:22",
          "sortId" : 1,
          "id" : "123456",
          "title" : "Rengöring"
        } ]
      }
    },
...
    "TodoInstantiate" : {
      "type" : "object",
      "required" : [ "assignedToId", "assignedToName", "assignedToObjectType", "dueAt" ],
      "properties" : {
        "assignedToId" : {
          "type" : "string",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "assignedToName" : {
          "type" : "string",
          "example" : "Frukt & Grönt"
        },
        "assignedToObjectType" : {
          "type" : "string",
          "example" : "Group"
        },
        "dueAt" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-04-24 14:41"
        }
      },
      "example" : {
        "dueAt" : "2018-04-24 14:41",
        "assignedToId" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
        "assignedToName" : "Frukt & Grönt",
        "assignedToObjectType" : "Group"
      }
    },
...
  },
...
}